+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To Be Released
 
+## 1.41.0
+
 Some features listed in this changelog, like databases next generation, are currently only available in preview to
 selected users.